Language.Thrift.Parser

Description

Provides a parser for Thrift IDLs.

In addition to parsing the IDLs, the parser also keeps track of Javadoc-style docstrings on defined items and makes their values available. For example,

/**
 * Fetches an item.
 */
Item getItem()

Note that the parser does not validate the Thrift file for correctness, so, for example, you could define a string value for an int constant.

Documentation

parseFromFile :: FilePath -> IO (Either (ParseError Char Dec) (Program SourcePos))

Parses the Thrift file at the given path.

parse :: (Stream s, Token s ~ Char) => FilePath -> s -> Either (ParseError Char Dec) (Program SourcePos)

parse name contents parses the contents of a Thrift document with name name held in contents.

thriftIDL :: (Stream s, Token s ~ Char) => Parsec Dec s (Program SourcePos)

Megaparsec parser that is able to parse full Thrift documents.

Components

program :: (Stream s, Token s ~ Char) => Parser s (Program SourcePos)

Top-level parser to parse complete Thrift documents.

header :: (Stream s, Token s ~ Char) => Parser s (Header SourcePos)

Headers defined for the IDL.

include :: (Stream s, Token s ~ Char) => Parser s (Include SourcePos)

The IDL includes another Thrift file.

include "common.thrift"

typedef common.Foo Bar

namespace :: (Stream s, Token s ~ Char) => Parser s (Namespace SourcePos)

Namespace directives allows control of the namespace or package name used by the generated code for certain languages.

namespace py my_service.generated

definition :: (Stream s, Token s ~ Char) => Parser s (Definition SourcePos)

A constant, type, or service definition.

constant :: (Stream s, Token s ~ Char) => Parser s (Const SourcePos)

A const definition.

const i32 code = 1;

typeDefinition :: (Stream s, Token s ~ Char) => Parser s (Type SourcePos)

A type definition.

service :: (Stream s, Token s ~ Char) => Parser s (Service SourcePos)

A service.

service MyService {
    // ...
}

typedef :: (Stream s, Token s ~ Char) => Parser s (Typedef SourcePos)

A typedef is just an alias for another type.

typedef common.Foo Bar

enum :: (Stream s, Token s ~ Char) => Parser s (Enum SourcePos)

Enums are sets of named integer values.

enum Role {
    User = 1, Admin

struct :: (Stream s, Token s ~ Char) => Parser s (Struct SourcePos)

A struct.

struct User {
    1: string name
    2: Role role = Role.User;
}

union :: (Stream s, Token s ~ Char) => Parser s (Union SourcePos)

A union of types.

union Value {
    1: string stringValue;
    2: i32 intValue;
}

exception :: (Stream s, Token s ~ Char) => Parser s (Exception SourcePos)

An exception that can be raised by service methods.

exception UserDoesNotExist {
    1: optional string message
    2: required string username
}

senum :: (Stream s, Token s ~ Char) => Parser s (Senum SourcePos)

An string-only enum. These are a deprecated feature of Thrift and shouldn't be used.

typeReference :: (Stream s, Token s ~ Char) => Parser s (TypeReference SourcePos)

A reference to a built-in or defined field.

constantValue :: (Stream s, Token s ~ Char) => Parser s (ConstValue SourcePos)

A constant value literal.

docstring :: (Stream s, Token s ~ Char) => Parser s Text

A javadoc-style docstring.

/**
 * foo
 */

This parses attempts to preserve indentation inside the docstring while getting rid of the aligned *s (if any) and any other preceding space.

Parser

type Parser s = StateT State (Parsec Dec s)

Underlying Parser type.

runParser :: (Stream s, Token s ~ Char) => Parser s a -> Parsec Dec s a

Evaluates the underlying parser with a default state and get the Megaparsec parser.

whiteSpace :: (Stream s, Token s ~ Char) => Parser s ()

Optional whitespace.