Language.Thrift.Pretty

Description

This module provides a pretty printer for Thrift IDLs. Most of the printers defined in this module produce output highlighted using ANSI escape codes. Get plain output by using plain.

Use prettyPrintHighlighted to produce output highlighted using ANSI escape codes. Note that this output will be unparseable and is suitable for printing inside a compatible terminal only. Use prettyPrint if you don't want highlighted output.

The behavior of the printer can be customized using Config objects.

The module also exports instances of the Pretty typeclass for elements of the AST.

Documentation

prettyPrintHighlighted :: Program ann -> Doc

Top-level pretty printer for Thrift documents.

prettyPrint :: Program ann -> Doc

Top-level pretty printer for Thrift documents that uses the default configuration (defaultConfig) for pretty printing.

Components

program :: Config -> Program ann -> Doc

Pretty print a Thrift IDL.

header :: Header ann -> Doc

Print the headers for a program.

include :: Include ann -> Doc

namespace :: Namespace ann -> Doc

definition :: Config -> Definition ann -> Doc

Print a constant, type, or service definition.

constant :: Config -> Const ann -> Doc

typeDefinition :: Config -> Type ann -> Doc

service :: Config -> Service ann -> Doc

typedef :: Config -> Typedef ann -> Doc

enum :: Config -> Enum ann -> Doc

struct :: Config -> Struct ann -> Doc

union :: Config -> Union ann -> Doc

exception :: Config -> Exception ann -> Doc

senum :: Config -> Senum ann -> Doc

typeReference :: Config -> TypeReference ann -> Doc

Pretty print a field type.

constantValue :: Config -> ConstValue ann -> Doc

Pretty print a constant value.

docstring :: Text -> Doc

Configuration

data Config

Configuration for the pretty printer.

Constructors

Config

Fields

indentWidth :: Int

Number of spaces to use for indentation.

defaultConfig :: Config

Default pretty printing configuration.