Language.Thrift.Pretty

Description

This module provides a pretty printer for Thrift IDLs. The pretty printer prserves docstrings specified for types.

The specifics of the printer can be configured using Config objects.

Documentation

prettyPrint :: Program ann -> Doc

Top-level pretty printer for Thrift documents that uses the default configuration (defaultConfig) for pretty printing.

Components

program :: Config -> Program ann -> Doc

Pretty print a Thrift IDL.

header :: Header -> Doc

Print the headers for a program.

definition :: Config -> Definition ann -> Doc

Print a constant, type, or service definition.

function :: Config -> Function ann -> Doc

Pretty print a function definition.

fieldType :: Config -> FieldType -> Doc

Pretty print a field type.

constantValue :: Config -> ConstValue -> Doc

Pretty print a constant value.

Configuration

data Config

Configuration for the pretty printer.

Constructors

Config

Fields

indentWidth :: Int

Number of spaces to use for indentation.

defaultConfig :: Config

Default pretty printing configuration.