Language.Thrift.Types

Description

This module defines types that compose a Thrift IDL file.

Most of the types have an optional srcAnnot parameter that represents a parser-specific source annotation. With trifecta this can hold the Delta. If you do not have need of this, you can use () as the parameter.

AST

data Program srcAnnot

A program represents a single Thrift document.

Constructors

Program

Fields

programHeaders :: [Header srcAnnot]

Headers in a document define includes and namespaces.

programDefinitions :: [Definition srcAnnot]

Types and services defined in the document.

headers :: Lens' (Program a) [Header a]

definitions :: Lens' (Program a) [Definition a]

data Header srcAnnot

Headers for a program.

Constructors

HeaderInclude (Include srcAnnot)
Request to include another Thrift file.

HeaderNamespace (Namespace srcAnnot)
A namespace specifier.

_Include :: Prism' (Header ann) (Include ann)

_Namespace :: Prism' (Header ann) (Namespace ann)

data Include srcAnnot

The IDL includes another Thrift file.

include "common.thrift"

typedef common.Foo Bar

Constructors

Include

Fields

includePath :: Text

Path to the included file.

includeSrcAnnot :: srcAnnot

path :: Lens' (Include a) Text

data Namespace srcAnnot

Namespace directives allows control of the namespace or package name used by the generated code for certain languages.

namespace py my_service.generated

Constructors

Namespace

Fields

namespaceLanguage :: Text

The language for which the namespace is being specified. This may be * to refer to all languages.

namespaceName :: Text

Namespace or package path to use in the generated code for that language.

namespaceSrcAnnot :: srcAnnot

language :: Lens' (Namespace a) Text

data Definition srcAnnot

A definition either consists of new constants, new types, or new services.

Constructors

ConstDefinition (Const srcAnnot)
A declared constant.

TypeDefinition (Type srcAnnot)
A custom type.

ServiceDefinition (Service srcAnnot)
A service definition.

_Const :: Prism' (Definition ann) (Const ann)

_Service :: Prism' (Definition ann) (Service ann)

_Type :: Prism' (Definition ann) (Type ann)

data Const srcAnnot

A declared constant.

const i32 code = 1;

Constructors

Const

Fields

constValueType :: TypeReference srcAnnot

Type of the constant.

constName :: Text

Name of the constant.

constValue :: ConstValue srcAnnot

Value of the constant.

constDocstring :: Docstring

Documentation.

constSrcAnnot :: srcAnnot

data Service srcAnnot

A service definition.

service MyService {
    // ...
}

Constructors

Service

Fields

serviceName :: Text

Name of the service.

serviceExtends :: Maybe Text

Name of the service this service extends.

serviceFunctions :: [Function srcAnnot]

All the functions defined for the service.

serviceAnnotations :: [TypeAnnotation]

Annotations added to the service.

serviceDocstring :: Docstring

Documentation.

serviceSrcAnnot :: srcAnnot

functions :: Lens' (Service a) [Function a]

extends :: Lens' (Service a) (Maybe Text)

data Type srcAnnot

Defines the various types that can be declared in Thrift.

Constructors

TypedefType (Typedef srcAnnot)
typedef

EnumType (Enum srcAnnot)
enum

StructType (Struct srcAnnot)
struct

UnionType (Union srcAnnot)
union

ExceptionType (Exception srcAnnot)
exception

SenumType (Senum srcAnnot)
senum

_Typedef :: Prism' (Type ann) (Typedef ann)

_Enum :: Prism' (Type ann) (Enum ann)

_Struct :: Prism' (Type ann) (Struct ann)

_Union :: Prism' (Type ann) (Union ann)

_Exception :: Prism' (Type ann) (Exception ann)

_Senum :: Prism' (Type ann) (Senum ann)

data Typedef srcAnnot

A typedef is just an alias for another type.

typedef common.Foo Bar

Constructors

Typedef

Fields

typedefTargetType :: TypeReference srcAnnot

The aliased type.

typedefName :: Text

Name of the new type.

typedefAnnotations :: [TypeAnnotation]

Annotations added to the typedef.

typedefDocstring :: Docstring

Documentation.

typedefSrcAnnot :: srcAnnot

targetType :: Lens' (Typedef a) (TypeReference a)

data Enum srcAnnot

Enums are sets of named integer values.

enum Role {
    User = 1, Admin = 2
}

Constructors

Enum

Fields

enumName :: Text

Name of the enum type.

enumValues :: [EnumDef srcAnnot]

Values defined in the enum.

enumAnnotations :: [TypeAnnotation]

Annotations added to the enum.

enumDocstring :: Docstring

Documentation.

enumSrcAnnot :: srcAnnot

data Struct srcAnnot

A struct definition

struct User {
    1: Role role = Role.User;
}

Constructors

Struct

Fields

structName :: Text

Name of the struct.

structFields :: [Field srcAnnot]

Fields defined in the struct.

structAnnotations :: [TypeAnnotation]

Annotations added to the struct.

structDocstring :: Docstring

Documentation.

structSrcAnnot :: srcAnnot

data Union srcAnnot

A union of other types.

union Value {
    1: string stringValue;
    2: i32 intValue;
}

Constructors

Union

Fields

unionName :: Text

Name of the union.

unionFields :: [Field srcAnnot]

Fields defined in the union.

unionAnnotations :: [TypeAnnotation]

Annotations added to the union.

unionDocstring :: Docstring

Documentation.

unionSrcAnnot :: srcAnnot

data Exception srcAnnot

Exception types.

exception UserDoesNotExist {
    1: optional string message
    2: required string username
}

Constructors

Exception

Fields

exceptionName :: Text

Name of the exception.

exceptionFields :: [Field srcAnnot]

Fields defined in the exception.

exceptionAnnotations :: [TypeAnnotation]

Annotations added to the exception.

exceptionDocstring :: Docstring

Documentation.

exceptionSrcAnnot :: srcAnnot

data Senum srcAnnot

An string-only enum. These are a deprecated feature of Thrift and shouldn't be used.

Constructors

Senum

Fields

senumName :: Text

senumValues :: [Text]

senumAnnotations :: [TypeAnnotation]

Annotations added to the senum.

senumDocstring :: Docstring

Documentation.

senumSrcAnnot :: srcAnnot

data FieldRequiredness

Whether a field is required or optional.

Constructors

Required
The field is required.

Optional
The field is optional.

_Required :: Prism' FieldRequiredness ()

_Optional :: Prism' FieldRequiredness ()

data Field srcAnnot

A field inside a struct, exception, or function parameters list.

Constructors

Field

Fields

fieldIdentifier :: Maybe Integer

Position of the field.

While this is optional, it is recommended that Thrift files always contain specific field IDs.

fieldRequiredness :: Maybe FieldRequiredness

Whether this field is required or not.

Behavior may differ between languages if requiredness is not specified. Therefore it's recommended that requiredness for a field is always specified.

fieldValueType :: TypeReference srcAnnot

Type of value the field holds.

fieldName :: Text

Name of the field.

fieldDefaultValue :: Maybe (ConstValue srcAnnot)

Default value of the field, if any.

fieldAnnotations :: [TypeAnnotation]

Field annotations.

fieldDocstring :: Docstring

Documentation.

fieldSrcAnnot :: srcAnnot

identifier :: Lens' (Field a) (Maybe Integer)

requiredness :: Lens' (Field a) (Maybe FieldRequiredness)

defaultValue :: Lens' (Field a) (Maybe (ConstValue a))

data EnumDef srcAnnot

A named value inside an enum.

Constructors

EnumDef

Fields

enumDefName :: Text

Name of the value.

enumDefValue :: Maybe Integer

Value attached to the enum for that name.

enumDefAnnotations :: [TypeAnnotation]

Annotations added to this enum field.

enumDefDocstring :: Docstring

Documentation

enumDefSrcAnnot :: srcAnnot

data ConstValue srcAnnot

A constant literal value in the IDL. Only a few basic types, lists, and maps can be presented in Thrift files as literals.

Constants are used for IDL-level constants and default values for fields.

Constructors

ConstInt Integer srcAnnot
An integer. 42

ConstFloat Double srcAnnot
A float. 4.2

ConstLiteral Text srcAnnot
A literal string. "hello"

ConstIdentifier Text srcAnnot
A reference to another constant. Foo.bar

ConstList [ConstValue srcAnnot] srcAnnot
A literal list containing other constant values. [42]

ConstMap [(ConstValue srcAnnot, ConstValue srcAnnot)] srcAnnot
A literal list containing other constant values. {"hellO": 1, "world": 2}

_ConstInt :: Prism' (ConstValue a) (Integer, a)

_ConstFloat :: Prism' (ConstValue a) (Double, a)

_ConstLiteral :: Prism' (ConstValue a) (Text, a)

_ConstIdentifier :: Prism' (ConstValue a) (Text, a)

_ConstList :: Prism' (ConstValue a) ([ConstValue a], a)

_ConstMap :: Prism' (ConstValue a) ([(ConstValue a, ConstValue a)], a)

data TypeReference srcAnnot

A reference to a type.

Constructors

DefinedType Text srcAnnot
A custom defined type referred to by name.

StringType [TypeAnnotation] srcAnnot
string and annotations.

BinaryType [TypeAnnotation] srcAnnot
binary and annotations.

SListType [TypeAnnotation] srcAnnot
slist and annotations.

BoolType [TypeAnnotation] srcAnnot
bool and annotations.

ByteType [TypeAnnotation] srcAnnot
byte and annotations.

I16Type [TypeAnnotation] srcAnnot
i16 and annotations.

I32Type [TypeAnnotation] srcAnnot
i32 and annotations.

I64Type [TypeAnnotation] srcAnnot
i64 and annotations.

DoubleType [TypeAnnotation] srcAnnot
double and annotations.

MapType (TypeReference srcAnnot) (TypeReference srcAnnot) [TypeAnnotation] srcAnnot
map<foo, bar> and annotations.

SetType (TypeReference srcAnnot) [TypeAnnotation] srcAnnot
set<baz> and annotations.

ListType (TypeReference srcAnnot) [TypeAnnotation] srcAnnot
list<qux> and annotations.

_DefinedType :: Prism' (TypeReference a) (Text, a)

_StringType :: Prism' (TypeReference a) ([TypeAnnotation], a)

_BinaryType :: Prism' (TypeReference a) ([TypeAnnotation], a)

_SListType :: Prism' (TypeReference a) ([TypeAnnotation], a)

_BoolType :: Prism' (TypeReference a) ([TypeAnnotation], a)

_ByteType :: Prism' (TypeReference a) ([TypeAnnotation], a)

_I16Type :: Prism' (TypeReference a) ([TypeAnnotation], a)

_I32Type :: Prism' (TypeReference a) ([TypeAnnotation], a)

_I64Type :: Prism' (TypeReference a) ([TypeAnnotation], a)

_DoubleType :: Prism' (TypeReference a) ([TypeAnnotation], a)

_MapType :: Prism' (TypeReference a) (TypeReference a, TypeReference a, [TypeAnnotation], a)

_SetType :: Prism' (TypeReference a) (TypeReference a, [TypeAnnotation], a)

_ListType :: Prism' (TypeReference a) (TypeReference a, [TypeAnnotation], a)

data Function srcAnnot

A function defined inside a service.

Constructors

Function

Fields

functionOneWay :: Bool

Whether the function is oneway. If it's one way, it cannot receive repsonses.

functionReturnType :: Maybe (TypeReference srcAnnot)

Return type of the function, or Nothing if it's void or oneway.

functionName :: Text

Name of the function.

functionParameters :: [Field srcAnnot]

Parameters accepted by the function.

functionExceptions :: Maybe [Field srcAnnot]

Exceptions raised by the function, if any.

functionAnnotations :: [TypeAnnotation]

Annotations added to the function.

functionDocstring :: Docstring

Documentation.

functionSrcAnnot :: srcAnnot

oneWay :: Lens' (Function a) Bool

returnType :: Lens' (Function a) (Maybe (TypeReference a))

parameters :: Lens' (Function a) [Field a]

exceptions :: Lens' (Function a) (Maybe [Field a])

data TypeAnnotation

Type annoations may be added in various places in the form,

(foo = "bar", baz, qux = "quux")

These do not usually affect code generation but allow for custom logic if writing your own code generator.

Constructors

TypeAnnotation

Fields

typeAnnotationName :: Text

Name of the annotation.

typeAnnotationValue :: Maybe Text

Value for the annotation.

type Docstring = Maybe Text

Docstrings are Javadoc-style comments attached various defined objects.

/**
 * Fetches an item.
 */
Item getItem()

Typeclasses

class HasAnnotations t where

Methods

annotations :: Lens' t [TypeAnnotation]

class HasDocstring t where

Methods

docstring :: Lens' t Docstring

class HasFields t where

Methods

fields :: Lens' (t a) [Field a]

class HasName t where

Methods

name :: Lens' t Text

class HasSrcAnnot t where

Methods

srcAnnot :: Lens' (t a) a

class HasValue s a | s -> a where

Methods

value :: Lens' s a

class HasValues s a | s -> a where

Methods

values :: Lens' s a

class HasValueType t where

Methods

valueType :: Lens' (t a) (TypeReference a)